requirements.txt
opencv-python
Cython
numba
progress
matplotlib
easydict
scipy
pyquaternion
nuscenes-devkit
pyyaml
motmetrics
scikit-learn
gs-wrap
gsutil
Cython
pytorch-lightning==1.1.8
tensorboard
tensorflow
gcsfs